Use sanitized production settings for test input
